Sync dev_be_hk files with dev_be_communal

+++ server/watchtower-api/Infrastructure/Repositories/ICommentRepository.cs
@@ -5,7 +5,7 @@ namespace WatchtowerApi.Infrastructure.Repositories;
 public interface ICommentRepository
 {
     Task<Comment> CreateAsync(int reportId, int userId, string commentText);
-    Task<Comment?> GetByIdAsync(long id);
+    Task<Comment?> GetByIdAsync(int id);
     Task<IEnumerable<Comment>> GetByReportIdAsync(int reportId);
     Task<IEnumerable<Comment>> GetByUserIdAsync(int userId);
     Task<Comment> UpdateAsync(int id, string commentText);
